% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleAutoComplete.R
\name{google_autocomplete}
\alias{google_autocomplete}
\title{Google Search Autocomplete}
\usage{
google_autocomplete(query, country = "com", searchLang = "lang_en",
  webInterfaceLang = "en")
}
\arguments{
\item{query}{The term for which autocomplete results are retrieved.}

\item{country}{Get only the country specific results not global results.}

\item{searchLang}{Language in which search is made.}

\item{webInterfaceLang}{Language of google search page.}
}
\value{
An list of terms based on the query.
}
\description{
This function is used to get google search auto complete suggestion.
You can get country specific google auto complete results and language
specific results.
}
\details{
\code{country} make sure that results retrieved for \code{query} are local
to mentioned country. \code{webInterfaceLang} parameter is for search
suggestions in specific language. \code{searchLang} is used to make search
in specific language. Please refer to \code{gSubDomain} data frame for
country availability and country codes. \code{gSearchLang} and
\code{gWebInLang} for search language and search web interface
availability respectively.
}
\note{
Too many requests can get you blocked. Have delay of 1-2 Sec for each request
}
\examples{
google_autocomplete("Where is olympics", "com","lang_en","en")

google_autocomplete("total athletes", "com")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
https://en.wikipedia.org/wiki/List_of_Google_domains
}
\seealso{
\code{\link{gSearchLang},\link{gSubDomain},\link{gWebInLang}}
}
\keyword{Auto}
\keyword{Complete}
\keyword{Google}
\keyword{Suggestions}

