% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAutoComplete.R
\name{googleAutoComplete}
\alias{googleAutoComplete}
\title{Google Auto Complete search suggestion results.}
\usage{
googleAutoComplete(query, country = "us", lang = "en")
}
\arguments{
\item{country}{The term on which search is performed}

\item{lang}{Language of google search page}
}
\value{
An list of terms which match the query
}
\description{
Get country specific google auto complete results. Country codes are avaliable in googleSubDomain
data set or wiki link[1]
}
\details{
\code{country} make sure that results retrieved for \code{query} are local to mentioned country.
\code{lang} parameter is for language specific search suggestions. Check the wikipedia link[1] for
google domains(Country Codes).
}
\examples{
wikiEdits("Auto Complete", "in","en")

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\references{
[1] https://en.wikipedia.org/wiki/List_of_Google_domains
}
\seealso{
\code{\link{}}
}
\keyword{Auto}
\keyword{Complete}
\keyword{Google}
\keyword{Suggestions}

