% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSearchLang-Data.R
\docType{data}
\name{googleSearchLang}
\alias{googleSearchLang}
\title{Goolge search language.}
\format{An data frame with 2 columns Language and Code.}
\source{
\href{#' https://sites.google.com/site/tomihasa/google-language-codes}
}
\usage{
data(googleSearchLang)
}
\description{
Goolge search language.
}
\examples{
data(googleSearchLang)
code <- googleSearchLang[googleSearchLang$Language == 'Japanese',2]
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
}
\keyword{datasets}

