% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleWebInterfaceLang-Data.R
\docType{data}
\name{googleWebInterfaceLang}
\alias{googleWebInterfaceLang}
\title{Goolge search web interface language.}
\format{An data frame with 2 columns Language and Code.}
\source{
\href{#' https://sites.google.com/site/tomihasa/google-language-codes}
}
\usage{
data(googleWebInterfaceLang)
}
\description{
Goolge search web interface language.
}
\examples{
data(googleWebInterfaceLang)
code <- googleWebInterfaceLang[googleWebInterfaceLang$Language == 'Hindi',2]
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
}
\keyword{datasets}

