% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSearchLang-Data.R
\docType{data}
\name{gSearchLang}
\alias{gSearchLang}
\title{Goolge search language.}
\format{An data frame with 2 columns Language and Code.}
\source{
\href{#' https://sites.google.com/site/tomihasa/google-language-codes}
}
\usage{
data(gSearchLang)
}
\description{
Goolge search language.
}
\examples{
data(gSearchLang)
code <- gSearchLang[gSearchLang$Language == 'Japanese',2]
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
}
\keyword{datasets}

