% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSubDomain-Data.R
\docType{data}
\name{gSubDomain}
\alias{gSubDomain}
\title{Goolge Search Domains}
\format{An data frame with 3 columns Region, Code and Domain.}
\source{
\href{https://en.wikipedia.org/wiki/List_of_Google_domains}
{Wikipedia Web Page}
}
\usage{
data(gSubDomain)
}
\description{
Goolge Search Domains
}
\examples{
data(gSubDomain)
regions <- gSubDomain$Region
}
\references{
https://en.wikipedia.org/wiki/List_of_Google_domains
}
\keyword{datasets}

