% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gWebInLang-Data.R
\docType{data}
\name{gWebInLang}
\alias{gWebInLang}
\title{Goolge search web interface language.}
\format{An data frame with 2 columns Language and Code.}
\source{
\href{#' https://sites.google.com/site/tomihasa/google-language-codes}
}
\usage{
data(gWebInLang)
}
\description{
Goolge search web interface language.
}
\examples{
data(gWebInLang)
code <- gWebInLang[gWebInLang$Language == 'Hindi',2]
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
}
\keyword{datasets}

