% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAutoComplete.R
\name{googleAutoComplete}
\alias{googleAutoComplete}
\title{Google Search Autocomplete}
\usage{
googleAutoComplete(query, country = "com", searchLang = "lang_en",
  webInterfacewebInterfaceLang = "en")
}
\arguments{
\item{query}{The term for which autocomplete results are retrieved.}

\item{country}{Get only the country specific results not global results.}

\item{searchLang}{Language in which search is made.}

\item{webInterfaceLang}{Language of google search page.}
}
\value{
An list of terms based on the query.
}
\description{
This function is used to get google search auto complete suggestion.
You can get country specific google auto complete results and language
specific results.
}
\details{
\code{country} make sure that results retrieved for \code{query} are local
to mentioned country. \code{webInterfaceLang} parameter is for search
suggestions in specific language. \code{searchLang} is used to make search
in specific language. Please refer to \code{googleSubDomain} data frame for
country availability and country codes. \code{googleSearchLang} and
\code{googleWebInterfaceLang} for search language and search web interface
availability respectively.
}
\note{
Please use it fair. Too many requests can get you blocked.
}
\examples{
googleAutoComplete("Where is olympics", "com","lang_en","en")
Results:
Where is olympics 2020
Where is olympics 2016
Where is olympics 2024
Where is olympics 2016 held

googleAutoComplete("total athletes", "com")
Results:
total athletes in 2014 Olympics
total athletes in 2014 winter Olympics
total athletes in the 2012 Olympics

}
\author{
Abhinav Yedla \email{abhinavyedla@gmail.com}
}
\references{
https://sites.google.com/site/tomihasa/google-language-codes
https://en.wikipedia.org/wiki/List_of_Google_domains
}
\seealso{
\code{\link{googleSearchLang},\link{googleWebInterfaceLang}}
}
\keyword{Auto}
\keyword{Complete}
\keyword{Google}
\keyword{Suggestions}

