% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSubDomain-Data.R
\docType{data}
\name{googleSubDomain}
\alias{googleSubDomain}
\title{Goolge Domain Data from taken from wikipedia page[1]}
\format{An data frame with 3 columns Region, TLD and Domain.}
\source{
\href{https://en.wikipedia.org/wiki/List_of_Google_domains}{Wikipedia Web Page}
}
\usage{
data(googleSubDomain)
}
\description{
Goolge Domain Data from taken from wikipedia page[1]
}
\examples{
data(googleSubDomain)
regions <- attr(googleSubDomain, "Region")
}
\references{

}
\keyword{datasets}

